+++ skew/resources/aws/ses.py
@@ -41,4 +41,4 @@ class Identity(AWSResource):
         return 'arn:aws:%s:%s:%s:%s/%s' % (
             self._client.service_name,
             self._client.region_name,
-            self._client.account_id, self.resourcetype, self.data)
+            self._client.account_id, self.resourcetype, self._data)